pyproject.toml
[project]
name = "guardian"
version = "1.0.0"
description = "Git Authentication & Development Assistant"
requires-python = ">=3.8"
dependencies = [ "markdown>=3.7", "markupsafe>=3.0.2", "beautifulsoup4>=4.12.0","aiohttp>=3.9.0", "pyyaml>=6.0.2", "secretstorage>=3.3.3", "astroid>=3.3.5", "autocommand>=2.2.2", "babel>=2.16.0", "backports.tarfile>=1.2.0", "certifi>=2024.8.30", "cffi>=1.17.1", "charset-normalizer>=3.4.0", "click>=8.1.7", "colorama>=0.4.6", "coverage>=7.6.4", "cryptography>=43.0.3", "dill>=0.3.9", "ghp-import>=2.1.0", "github-cli>=1.0.0", "idna>=3.10", "importlib-metadata>=8.5.0", "inflect>=7.4.0", "iniconfig>=2.0.0", "jaraco.classes>=3.4.0", "jaraco.collections>=5.1.0", "jaraco.context>=6.0.1", "jaraco.functools>=4.1.0", "jaraco.text>=4.0.0", "jeepney>=0.8.0", "jinja2>=3.1.4", "jwt>=1.3.1", "keyring>=25.5.0", "markdown-it-py>=3.0.0", "mccabe>=0.7.0", "mdurl>=0.1.2", "mergedeep>=1.3.4", "mitmproxy>=10.0.0","mkdocs>=1.6.1", "mkdocs-get-deps>=0.2.0", "mkdocs-material>=9.5.44", "mkdocs-material-extensions>=1.3.1", "more-itertools>=10.5.0", "mypy-extensions>=1.0.0", "packaging>=24.2", "paginate>=0.5.7", "pathspec>=0.12.1", "pip>=24.2", "platformdirs>=4.3.6", "pluggy>=1.5.0", "pycparser>=2.22", "pygments>=2.18.0", "pylint>=3.3.1", "pymdown-extensions>=10.12", "pytest-cov>=6.0.0", "python-dateutil>=2.9.0.post0", "pyyaml-env-tag>=0.1", "pyyaml>=6.0.1", "regex>=2024.11.6", "requests>=2.32.3", "rich>=13.9.4", "setuptools>=75.4.0", "simplejson>=3.19.3", "six>=1.16.0", "toml>=0.10.2", "tomli>=2.1.0", "tomlkit>=0.13.2", "typeguard>=4.4.1", "typing-extensions>=4.12.2", "urllib3>=2.2.3", "watchdog>=6.0.0", "wheel>=0.45.0", "zipp>=3.21.0",]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [ "black>=24.10.0", "isort>=5.13.2", "mypy>=1.13.0", "pytest>=8.3.3",]

[project.scripts]
guardian = "guardian.cli:cli"

[project.urls]
Homepage = "https://github.com/None/-home-thatch-dev-guardian"
Documentation = "https://-home-thatch-dev-guardian.readthedocs.io/"
Repository = "https://github.com/None/-home-thatch-dev-guardian.git"

[tool.black]
line-length = 88
target-version = [ "py38",]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = [ "tests",]
python_files = [ "test_*.py",]
addopts = "--cov=guardian --cov-report=xml --cov-report=term-missing"

[tool.hatch.build.targets.wheel]
packages = [ "src/guardian",]
